add mappedPermissions service

+++ common/common-custom-security/src/main/java/com/meysam/common/customsecurity/model/dto/PermissionDTO.java
@@ -19,7 +19,7 @@ public class PermissionDTO implements Serializable {
 
     private String enKey;
 
-    private int parent;
+    private Long parent;
 
 
 }
